requirements.txt
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.1
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.0.0
python-dotenv>=1.0.0
click>=8.0.0
requests>=2.28.0

